backend/requirements.txt
anyio==4.10.0
asgiref==3.9.1
autopep8==2.3.2
black==25.11.0
cachetools==6.2.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.2.0
django-appconf==1.1.0
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-cryptography==1.1
django-extensions==4.1
django-js-asset==3.1.2
django-ratelimit==4.1.0
django-social-auth==0.7.28
djangorestframework==3.16.1
git-filter-repo==2.47.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gunicorn==23.0.0
h11==0.16.0
httplib2==0.22.0
httptools==0.6.4
idna==3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
mysqlclient==2.2.7
oauth2==1.9.0.post1
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
proto-plus==1.26.1
protobuf==6.33.1
psycopg==3.2.11
psycopg-binary==3.2.11
psycopg-pool==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-decouple==3.8
python-dotenv==1.1.1
python-http-client==3.3.7
python-openid==2.2.5
pytokens==0.3.0
PyYAML==6.0.2
ratelimit==2.2.1
requests==2.32.5
requests-oauthlib==2.0.0
resend==2.19.0
rsa==4.9.1
sendgrid==6.12.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.4
whitenoise==6.10.0
python-decouple 3.8